⚡ BarkAI Local Implementation (#3)
* 🚧 - WIP Unstable: Checkpoint after pass config to env vars + refacto for Pylint. TODO: Get the work same/better as main branch without regression

* 🧹 - Pylint 10/10

* 🧹 - Freeze requirements + keep only 3.10.11 version in pipelines

* 🐞 - Debugging pipeline

* 🐞 - Debugging pipeline

* 🐞 - Debugging pipeline

* 🐞 - Disable Pylint pipeline ATM because some package don't work on their servers

* ⚡  - Fix BAarkAI implementation
