+++ README.md
@@ -16,9 +16,35 @@ If you have any questions about this repo or would like to share your ideas for 
         <img src="https://img.shields.io/badge/pytorch-EE4C2C?style=for-the-badge&logo=pytorch&logoColor=white" />
 </div>
 
+## 📖 Table of Contents
+
+- [🤖 Ava.py](#🤖-avapy)
+  - [📖 Table of Contents](#📖-table-of-contents)
+  - [🛠️ Prerequisites](#️🛠️-prerequisites)
+  - [📦 Installation](#📦-installation)
+
 ## 🛠️ Prerequisites
 
-**Work in progress**
+- [Python 3.10](https://www.python.org/) (**recommended** *not tested on other versions*)
+- [Anaconda](https://www.anaconda.com/) (**recommended** *for a better experience with virtual environments*)
+- [Nvidia CUDA](https://developer.nvidia.com/cuda-toolkit) (*if you want to use GPU for local models*)
+- [An OpenAI API key](https://beta.openai.com/) (*if you want to use OpenAI API*)
+
+## 🎨 Features
+
+#### Audio detection & recording
+- pyaudio (Local)
+#### Speech recognition
+- OpenAI Whisper (Local) *⚡Need a good GPU for large model, good CPU for small model*
+- OpenAI Whisper (API)
+
+#### Text-to-speech
+- System TTS (pyttsx3)
+- Bark AI (Local) *⚡Need a very good GPU*
+- Eleven Labs (API) *⌛ Not yet implemented*
+#### Text generation
+- OpenAI GPT-3.5-Turbo (API)
+- Falcon-40B (Local) *⚡Need a very good GPU*
 
 ## 📦 Installation
