+++ docs/.env.example
@@ -1,40 +1,40 @@
-IS_DEV="True"
+IS_DEV=True
 
-OPENAI_API_KEY="sk-<your-openai-api-key>"
-OPENAI_LANGUAGE="fr"
-OPENAI_WHISPER_ENGINE="whisper-1"
-OPENAI_GPT_ENGINE="gpt-3.5-turbo"
+OPENAI_API_KEY=sk-...
+OPENAI_LANGUAGE=fr
+OPENAI_WHISPER_ENGINE=whisper-1
+OPENAI_GPT_ENGINE=gpt-3.5-turbo
 
-CHATBOT_USE_LOCAL="False"
-CHATBOT_MAX_LENGTH="300"
-CHATBOT_LOCAL_DEVICE_MAP="auto"
-CHATBOT_LOCAL_MODEL="tiiuae/falcon-7b-instruct"
-CHATBOT_LOCAL_PIPELINE="text-generation"
-CHATBOT_LOCAL_CACHE_DIR="./models/"
-CHATBOT_LOCAL_TOP_K="10"
-CHATBOT_LOCAL_NUM_RETURN_SEQUENCES="1"
-CHATBOT_LOCAL_DO_SAMPLE="True"
-CHATBOT_LOCAL_TRUST_REMOTE_CODE="True"
+CHATBOT_USE_LOCAL=0
+CHATBOT_MAX_LENGTH=300
+CHATBOT_LOCAL_DEVICE_MAP=auto
+CHATBOT_LOCAL_MODEL=tiiuae/falcon-7b-instruct
+CHATBOT_LOCAL_PIPELINE=text-generation
+CHATBOT_LOCAL_CACHE_DIR=./models/
+CHATBOT_LOCAL_TOP_K=10
+CHATBOT_LOCAL_NUM_RETURN_SEQUENCES=1
+CHATBOT_LOCAL_DO_SAMPLE=True
+CHATBOT_LOCAL_TRUST_REMOTE_CODE=True
 
-RECORDER_DIRECTORY="./records/"
-RECORDER_CHUNK="1024"
-RECORDER_CHANNELS="1"
-RECORDER_RATE="44100"
-RECORDER_SWIDTH="2"
-RECORDER_TIMEOUT_LENGTH="1"
-RECORDER_RMS_THRESHOLD="10"
-RECORDER_INPUT="True"
+RECORDER_DIRECTORY=./records/
+RECORDER_CHUNK=1024
+RECORDER_CHANNELS=1
+RECORDER_RATE=44100
+RECORDER_SWIDTH=2
+RECORDER_TIMEOUT_LENGTH=1
+RECORDER_RMS_THRESHOLD=10
+RECORDER_INPUT=True
 
-TTS_USE_LOCAL="False"
-TTS_LANGUAGE="fr"
-TTS_LOCAL_SPEAKER="v2/fr_speaker_1"
-TTS_LOCAL_GENERATION_TEMP="0.6"
-TTS_LOCAL_MIN_EOS_P="0.05"
-TTS_LOCAL_CUDA_VISIBLE_DEVICES="0"
-TTS_LOCAL_OFFLOAD_CPU="False"
-TTS_LOCAL_USE_SMALL_MODELS="True"
+TTS_USE_LOCAL=0
+TTS_LANGUAGE=fr
+TTS_LOCAL_SPEAKER=v2/fr_speaker_1
+TTS_LOCAL_GENERATION_TEMP=0.6
+TTS_LOCAL_MIN_EOS_P=0.05
+TTS_LOCAL_CUDA_VISIBLE_DEVICES=0
+TTS_LOCAL_OFFLOAD_CPU=False
+TTS_LOCAL_USE_SMALL_MODELS=True
 
-SPEECH_RECOGNITION_USE_LOCAL="True"
-SPEECH_RECOGNITION_LANGUAGE="fr"
-SPEECH_RECOGNITION_LOCAL_DEVICE="cuda"
-SPEECH_RECOGNITION_LOCAL_MODEL="medium"
+SPEECH_RECOGNITION_USE_LOCAL=1
+SPEECH_RECOGNITION_LANGUAGE=fr
+SPEECH_RECOGNITION_LOCAL_DEVICE=cuda
+SPEECH_RECOGNITION_LOCAL_MODEL=medium
